+++ src/Normalizer/VariableFromConfigMapNormalizer.php
@@ -50,19 +50,25 @@ class VariableFromConfigMapNormalizer implements DenormalizerInterface, Normaliz
         } elseif (\array_key_exists('configMap', $data) && null === $data['configMap']) {
             $object->setConfigMap(null);
         }
-        if (\array_key_exists('item', $data)) {
+        if (\array_key_exists('item', $data) && null !== $data['item']) {
             $object->setItem($data['item']);
+        } elseif (\array_key_exists('item', $data) && null === $data['item']) {
+            $object->setItem(null);
         }
-        if (\array_key_exists('id', $data)) {
+        if (\array_key_exists('id', $data) && null !== $data['id']) {
             $object->setId($data['id']);
+        } elseif (\array_key_exists('id', $data) && null === $data['id']) {
+            $object->setId(null);
         }
         if (\array_key_exists('environment', $data) && null !== $data['environment']) {
             $object->setEnvironment($data['environment']);
         } elseif (\array_key_exists('environment', $data) && null === $data['environment']) {
             $object->setEnvironment(null);
         }
-        if (\array_key_exists('name', $data)) {
+        if (\array_key_exists('name', $data) && null !== $data['name']) {
             $object->setName($data['name']);
+        } elseif (\array_key_exists('name', $data) && null === $data['name']) {
+            $object->setName(null);
         }
 
         return $object;
